Update chunk data packet cache lazily
Previously, the chunk data cache was regenerated every time a block or
biome was modified in a chunk. This causes performance problems in
situations where chunks are being modified every tick, like the conway
example or redstone machines.

This changes things so that the chunk data packet is regenerated only
when a client actually needs to load the chunk. This isn't perfect
beacause the regeneration cannot happen in parallel, but the benefits
outweigh the costs.

This has revealed that compression is surprisingly slow. This
should be investigated later.