[ECS] Add Inventory component, re-implement player inventories (#187)
- add missing field preventing build, format
- add Inventory ECS component
- add cursor_item fields, getters, setters to Client
- add new_with_title to Inventory instead of using an Option
- add Inventory component to client entities when they connect
- add system update_player_inventories

This does not implement opening other inventories yet, planning to do
that in a follow up PR.

The idea here is to have an `Inventory` component that can be used for
all inventories. If you want a chest to have an inventory, give the
corresponding block entity an `Inventory`. If you don't want any
inventory update systems to run, simply don't give clients a player
inventory.