Update Player List (#195)
The player list has been made into a resource rather than a component
for the sake of simplicity. If we actually need to support multiple
simultaneous player lists, we should be able to add that without too
much hassle.

`cargo r -p valence_new --example player_list` for the example.