Add `ChunkView` abstraction (#213)
A chunk view is a view distance combined with a center `ChunkPos`. It
represents the set of chunks that a client can see.

`ChunkView` simplifies some code internally as well as making some
examples like `parkour` and `terrain` easier to implement.

The first commit has introduced a bug that I am looking for. Go to the
`cow_sphere` example, set your render distance to 2, and then fly away
from and towards the sphere. Eventually there will be a giant hole in
the sphere.

I might need to set up unit tests for the chunk/entity code first.

- [x] Find and fix bug.