Chunk Rework (#402)
## Description

Iterate on the design of `valence_instance` and `valence_anvil`. This
has been a blocker for #193 and #342.

- Removed the `LOADED` const param from `Chunk` and split it into
separate `LoadedChunk` and `UnloadedChunk` types. Common operations
between the two are part of the new `Chunk` trait.
- Reworked the chunk interface and made it more convenient.
- Removed the confusing `ChunkCell` mechanism and flattened its data
into `LoadedChunk`.
- As a consequence of the previous point, `LoadedChunk` now implements
`WritePacket`.
- Added support for the new biome change packet introduced circa 1.19.4.
- Resolved some confusion around the client's respawn position/compass
position. Fixes #312.
- Improvements to `valence_anvil`. Now handles chunks with unusual
`min_y` correctly.
- Simplified `PacketWriter` interface. No longer takes a mut ref to a
`Vec<u8>`.
- Fixed a bug where block entities were not properly removed when
truncating a chunk.
- Added more unit tests for chunk mutations and chunk loading/unloading.
- Loaded chunks now use viewership information for optimization. Only
chunks in view of clients will record changes. For instance, writing a
packet to a chunk out of view is a no-op.
- Entities are always despawned on the client when the entity is moved
to an unloaded chunk position.
- Reorganized `valence_instance`.
- Fix player list systems not being added to the correct system set.
- New utilities for unit tests.

There was initially some indecision around whether chunks and block
entities should be entities in the ECS. After some consideration, I
decided to abandon that idea.