Sorted Chunk View (#403)
## Description

Change `ChunkView` so that iteration is done in sorted order by the
distance to the center of the view. This is accomplished by precomputing
a lookup table at compile time for all the possible view distances.

As a result of this change, the most important packets for clients to
receive will typically arrive first. This is especially important when
clients are loading into a new area, since the chunks closest to the
client ought to be loaded first.