Reorganize Crates (#449)
# Objective

Once again I am trying to reorganize the crate structure into something
I'm happy with. Hopefully I got it right this time.

The goal is to make it easier to extend valence without an explosion of
tiny crates or other problems. Also fixing miscellaneous issues along
the way.

# Solution

- Replaced `valence_client` with `valence_server` to accommodate a wider
variety of modules. `valence_server` is glob re-exported from the main
`valence` crate.
- All of the modules that were previously in `valence_client` have
become individual plugins for more flexibility.
- Merged `valence_server` and `valence_layer` since they were already
tightly coupled. Removed lots of `#[doc(hidden)]` as a result.
- Brought back `valence_protocol` to contain all protocol related stuff.
Replaces `valence_packet` and `valence_core`. This can be used by
servers, clients, proxies, and other tools without dragging in
valence-specific things.
- Consolidated most generated code into `valence_generated`. Keeping
this in one crate resolves some potential circular dependency issues
between things like items and blocks. `valence_protocol` depends on this
so we can use `BlockState` and other types in packets. It might slow
down the build too much, so we could split it up if we have to.
- Created `valence_math`. This re-exports `glam` + our own `Aabb` type
which was previously in `valence_core`.
- Merged `valence_biome` and `valence_dimension` into
`valence_registry`.
- Split off JSON text and resource identifiers into `valence_text` and
`valence_ident` respectively.
- Created `valence_lang` to hold language related code like translation
keys. There's not much there now, but it could be used for a registry of
languages and translation keys in the future.
- Added type alias for the compression threshold type.
- Cleaned up many imports.

![depgraph](https://raw.githubusercontent.com/valence-rs/valence/1836693e5be47cdfc854c5734be2ae7ca4dfbaa3/assets/depgraph.svg)