Added food and saturation (#568)
# Objective

- Food and saturation currently don't exist for the player in valence
- Closes #565 

# Solution

Adds food and saturation to the `PlayerEntity` and tells the client
about any changes. The only problem right now is that the default value
for `Health` is the same for ALL entities, so the server tells the
client that their health is at 14 when it should be at 20.

## Note

Uncertain if best approach. Please lmk if another approach would be
better