Fix command argument parse error checking (#582)
# Objective

This pull request aims to fix an error when parsing commands with
arguments. This error can be recreated by running the command example
and executing the command "/gamemode creative @a" or "/gamemode creative
{username}" etc. in game.

# Solution

The order of subtraction between the pre-parse argument string and
post-parse argument string was changed to hopefully accomplish the
intended purpose. This fixes the error when running the command in the
example.

Note: I'm not an expert on how command parsing works, maybe
@JackCrumpLeys can verify the solution?